+++ src/intervention_model/AttributeTable.h
@@ -4,8 +4,8 @@
   * SPDX-License-Identifier: MIT
   */
 
-#ifndef INTERVENTION_MODEL_ATTRIBUTETABLE_H_
-#define INTERVENTION_MODEL_ATTRIBUTETABLE_H_
+#ifndef ATTRIBUTETABLE_H_
+#define ATTRIBUTETABLE_H_
 
 #include "../readers/DataInterface.h"
 #include "../readers/DataReader.h"
@@ -34,7 +34,6 @@ class AttributeTable {
   explicit AttributeTable(bool isPersonTable);
   Attribute getAttribute(int i);
   union Data getDefaultValue(int i) const;
-  double getDefaultValueAsDouble(int i) const;
   std::string getName(int i) const;
   DataTypes::DataType getDataType(int i);
   bool getTableType();
@@ -46,4 +45,4 @@ class AttributeTable {
   void readAttributes(const AttributeList &dataFormat);
 };
 
-#endif  // INTERVENTION_MODEL_ATTRIBUTETABLE_H_
+#endif  // ATTRIBUTETABLE_H_
