+++ src/intervention_model/Interventions.cpp
@@ -6,8 +6,8 @@
 
 #include "Interventions.h"
 #include "AttributeTable.h"
-#include "../protobuf/interventions.pb.h"
 #include "../readers/DataInterface.h"
+#include "../readers/interventions.pb.h"
 
 #include <vector>
 
@@ -26,15 +26,12 @@ VaccinationIntervention::VaccinationIntervention(
   vaccinationProbability = interventionDef.vaccination().probability();
   vaccinatedSusceptibility = interventionDef.vaccination()
     .vaccinated_susceptibility();
-  CkPrintf("Vaccination: prob: %f, susceptibility: %f\n",
-      vaccinationProbability, vaccinatedSusceptibility);
   this->vaccinatedIndex = t.getAttribute("vaccinated");
   this->susceptibilityIndex = t.getAttribute("susceptibility");
 }
 
 void VaccinationIntervention::pup(PUP::er &p) {
   p | vaccinationProbability;
-  p | vaccinatedSusceptibility;
   p | vaccinatedIndex;
   p | susceptibilityIndex;
 }
@@ -43,6 +40,9 @@ bool VaccinationIntervention::test(const DataInterface &p,
     std::default_random_engine *generator) const {
   std::uniform_real_distribution<double> distribution(0.0, 1.0);
 
+  // CkPrintf("Vaccinated:%d, riskProbability: %f, selected for vaccine %d\n",
+  //     p.getValue(vaccinatedIndex).boolean, p.getValue(riskIndex).probability,
+  //     applied);
   return !p.getValue(vaccinatedIndex).boolean
     && distribution(*generator) < vaccinationProbability;
 }
